Updates to clean up warning in Visual Studio.  This is the first of several commits.  This commit cleans up all of the "defined but not used" messages for exceptions, and adds code so that the export warning for std::string and (most) std::vector classes are no longer generated.  std::map, std::list, and std::ifstream will still generate messages.

+++ src/base/spacecraft/TextTrajectoryFile.cpp
@@ -93,9 +93,9 @@ bool TextTrajectoryFile::ReadLine()
         mInStream >> mTempData[i];
 
     mTrajData.time = mTempData[0];
-    mTrajData.x = mTempData[1];
-    mTrajData.y = mTempData[2];
-    mTrajData.z = mTempData[3];
+    mTrajData.x = (float)mTempData[1];
+    mTrajData.y = (float)mTempData[2];
+    mTrajData.z = (float)mTempData[3];
     
     // add to mData
     mDataArray.push_back(mTrajData);
